// src/bot/types/enum.ts
export enum Languages {
  UZ = "uz",
  RU = "ru",
  // ENG = 'eng'
}

export enum CallbackActions {
  DISABLED = "disabled",
  MAIN_MENU = "main_menu_cb",
  ORDERS = "orders_cb",
  ABOUT = "about_cb",
  SETTINGS = "settings_cb",
  LANG_LIST = "lang_list_cb",
  CHANGE_LANG = "change_lang_cb",
  PRODUCT_CALLBACK = "product",
  OPERATOR = "operator_cb",
  ADMIN_UPLOAD_CODES = "admin_upload_codes",
  ADMIN_UPLOAD_WINNERS = "admin_upload_winners",
  ADMIN_CLEAR_CODES = "admin_clear_codes",
  ADMIN_CLEAR_WINNERS = "admin_clear_winners",
  ADMIN_UPLOAD_IMAGES = "admin_upload_images",
}

export enum Operator {
  INCREMENT = "inc",
  DECREMENT = "dec",
  Addition = "Addition",
  Subtraction = "Subtraction",
  Multiplication = "Multiplication",
  Exponentiation = "Exponentiation",
  Division = "Division",
  Modulus = "Modulus",
}